Update Shifts And Relatives

+++ app/Http/Controllers/ShiftsController.php
@@ -20,8 +20,8 @@ class ShiftsController extends Controller
         $input = $request->validate([
             'shift_id' => "required|string",
             'shift_name' => "required|string",
-            'start_time' => "required|time",
-            'end_time' => "required|time",
+            'start_time' => "required|date_format:H:i",
+            'end_time' => "required|date_format:H:i",
         ]);
         Shifts::create($input);
         return response()->json([], 201);
@@ -34,12 +34,12 @@ class ShiftsController extends Controller
 
     public function update(Request $request)
     {
-        $checkOut = Shifts::find($request->timekeeping_id);
+        $checkOut = Shifts::find($request->shift_id);
         $input = $request->validate([
             'shift_id' => "required|string",
             'shift_name' => "required|string",
-            'start_time' => "required|time",
-            'end_time' => "required|time",
+            'start_time' => "required|date_format:H:i",
+            'end_time' => "required|date_format:H:i",
         ]);
         $checkOut->shift_id = $input['shift_id'];
         $checkOut->shift_name = $input['shift_name'];
@@ -49,9 +49,28 @@ class ShiftsController extends Controller
         return response()->json([], 200);
     }
 
-    public function delete(Shifts $shifts)
+    // public function delete(Shifts $shifts)
+    // {
+    //     $shifts->delete();
+    //     return response()->json(["message" => "Delete success",], 200);
+    // }
+    public function delete($id)
     {
+        $shifts = Shifts::find($id);
+    
+        if (!$shifts) {
+            return response()->json([
+                "status" => false,
+                "message" => "Shifts not found",
+                "data" => []
+            ], 404);
+        }
+    
         $shifts->delete();
-        return response()->json(["message" => "Delete success",], 200);
+        return response()->json([
+            "status" => true,
+            "message" => "Delete success",
+            "data" => []
+        ], 200);
     }
 }
